Added collision 0.5, works. . .
There is a bug in it that if you move downright at a certain speed I
think you can move through walls, I have had a hard time replicating it
though. It is not fluid either. That needs fixing. I am thinking of
asking on stack overflow or some gamedev site for advice. This is hard.